+++ src/main/java/com/renomad/minum/logging/LoggingActionQueue.java
@@ -23,6 +23,10 @@ final class LoggingActionQueue implements AbstractActionQueue {
     private final ExecutorService executorService;
     private final LinkedBlockingQueue<RunnableWithDescription> queue;
     private Thread queueThread;
+    /**
+     * Set as volatile so that multiple threads may see the update
+     * as soon as it occurs.
+     */
     private volatile boolean stop = false;
     private boolean isStoppedStatus = false;
     private final Map<LoggingLevel, Boolean> enabledLogLevels;