+++ src/main/java/com/renomad/minum/logging/LoggingLevel.java
@@ -11,10 +11,9 @@ public enum LoggingLevel {
     DEBUG,
 
     /**
-     * This level is used when we catch an {@link Error} or {@link java.io.IOException}
-     * that has bubbled up to the top of the call stack.  It reflects situations that
-     * are cannot be reasonably handled by the developer, such as bugs in the operating
-     * system and hardware failures.
+     * This level is used when we catch an {@link Exception} that has bubbled up to
+     * the top of the call stack.  It reflects situations that cannot be easily
+     * anticipated by the developer, such as bugs in the operating system or hardware failures.
      */
     WARN,