+++ src/test/java/com/renomad/minum/utils/CryptoUtilsTests.java
@@ -11,22 +11,20 @@ public class CryptoUtilsTests {
 
     @Test
     public void testBytesToHex() {
-        var cryptoUtils = new CryptoUtils();
-        assertEquals(cryptoUtils.bytesToHex(new byte[]{1, 2, 3}), "010203");
-        assertEquals(cryptoUtils.bytesToHex(new byte[0]), "");
-        assertEquals(cryptoUtils.bytesToHex(new byte[]{0, (byte) 255,(byte) 255,(byte) 255}), "00ffffff");
+        assertEquals(CryptoUtils.bytesToHex(new byte[]{1, 2, 3}), "010203");
+        assertEquals(CryptoUtils.bytesToHex(new byte[0]), "");
+        assertEquals(CryptoUtils.bytesToHex(new byte[]{0, (byte) 255,(byte) 255,(byte) 255}), "00ffffff");
     }
 
     @Test
     public void testCreatePasswordHash_BadAlgorithm() {
-        var cryptoUtils = new CryptoUtils();
-        var ex = assertThrows(NoSuchAlgorithmException.class, () -> CryptoUtils.createPasswordHash("foo_password", "mysalt", "badalgorithm"));
-        assertEquals(ex.getMessage(), "badalgorithm SecretKeyFactory not available");
+        var ex = assertThrows(UtilsException.class, () -> CryptoUtils.createPasswordHash("foo_password", "mysalt", "badalgorithm"));
+        assertEquals(ex.getMessage(), "java.security.NoSuchAlgorithmException: badalgorithm SecretKeyFactory not available");
+        assertEquals(ex.getCause().getMessage(), "badalgorithm SecretKeyFactory not available");
     }
 
     @Test
     public void testCreatePasswordHash() {
-        var cryptoUtils = new CryptoUtils();
         String passwordHash = CryptoUtils.createPasswordHash("abc123", "saltysalt");
         assertEquals(passwordHash, "87f83512f4c18af25e82728cfdf96194");
     }